+++ src/scenes/AcceleratorRouter/ParticipantProjects/ProjectDocumentsView.tsx
@@ -1,0 +1,28 @@
+import React from 'react';
+
+import { useTheme } from '@mui/material';
+
+import Card from 'src/components/common/Card';
+
+import DocumentsTable from '../Documents/DocumentsView/DocumentsTable';
+
+type ProjectDocumentsViewProps = {
+  projectId: number;
+};
+
+const ProjectDocumentsView = ({ projectId }: ProjectDocumentsViewProps) => {
+  const theme = useTheme();
+
+  return (
+    <Card
+      style={{
+        borderRadius: theme.spacing(1),
+        padding: 0,
+      }}
+    >
+      <DocumentsTable showTitle />
+    </Card>
+  );
+};
+
+export default ProjectDocumentsView;
